Fixes dtests failures with the new token allocation algorithm (CASSANDRA-13701).
ccm has methods which allow to wait for some event to happen, like node startup, which use a fixed timeout.
This PR makes that timeout a constant instead of being harcoded and extends it from 30s to 45s to account for the extra wait that the new token allocation algorithm includes to wait for gossip state to settle.
Force nodes to start sequentially when starting the whole cluster